requirements.txt
pillow>=12.0.0
numpy>=2.0.0
matplotlib>=3.7.1
pandas>=2.0.0
Anthropic
openai
xai_sdk
pydantic
boto3
scipy
google
skyfield
google.genai
